pyproject.toml
[project]
name = "autonomous-research-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.3.1",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "biopython>=1.86",
    "boto3>=1.42.4",
    "chonkie>=1.4.2",
    "einops>=0.8.1",
    "fastapi>=0.115.0,<0.124.0",
    "google-adk>=1.14.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "langchain>=1.1.2",
    "langgraph>=1.0.4",
    "lxml>=6.0.2",
    "neo4j>=6.0.3",
    "nomic>=3.9.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "prometheus-client>=0.23.1",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.4.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "redis>=7.1.0",
    "sentence-transformers>=5.1.2",
    "spacy>=3.8.11",
    "sqlalchemy>=2.0.44",
    "streamlit>=1.52.1",
    "trafilatura>=2.0.0",
    "unstructured>=0.18.21",
    "uvicorn>=0.38.0",
    "wikipedia-api>=0.8.1",
]

[dependency-groups]
dev = [
    "mypy>=1.19.0",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.8",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"


